Fix dependency of odk.py on template dir
Before, the odk script referred to the local volume root for finding the template directory, which only worked if the template directory was actually there (basically, you needed to clone the odk gh repo and run the seed script from in there, https://github.com/INCATools/ontology-development-kit/issues/209). Now, the odk.py refers to the internal template dir and the tests pass.