Create dummy Git config to run the tests.
Running the test suite requires a minimal Git configuration. The
seed-via-docker.sh script normally takes care of that by mapping the
calling user's own $HOME/.gitconfig file inside the Docker container
used to run the tests, but this assumes that the calling user does have
a .gitconfig file -- an assumption that is seemingly wrong on the
machines used to run the Github Actions workflows.