Fix installation of Konclude.
Konclude is broken an all platforms: on x86_64, the version we used to
install depends on the musl C library that is not provided in the image;
on arm64, we don't have a compatible binary.

On x86_64, we now install a more recent version of Konclude that depends
on the standard GNU libc. On arm64, we compile Konclude ourselves.