Add a --platform argument when calling docker build.
Docker should default to build an image targeting the architecture it is
currently running on, but for some reason on M1 that does not seem to
always work and it may sometimes try to build an amd64 image instead
(thus causing all sorts of failure down the line). Better be explicit
and say which architecture we are targeting.

The default is taken from the OS system's uname command, so it *should*
always give the expected result without any user intervention. If
needed, the user can override the automatic detection of architecture by
assigning the ARCH variable herself when calling make.