Update multi-arch builds.
Make sure to build the builder image (as a multi-arch image itself)
before attempting to build the final multi-arch images.

Because it does not seem there is a way to load a multi-arch image
locally, the builder image will have to be pushed to the Docker Hub
before it can be used to build the final odklite and odkfull images.
This is silly and inefficient, but Docker does not offer any other
solution.