Avoid installing Python packages twice.
When installing the Python packages for the odkfull image, pip is
ignorant of the packages we have already installed for the odklite
image, so some packages already installed in the odklite staging tree
may be installed again in the odkfull staging tree, if they are
dependencies of the odkfull packages. This would result in artificially
and needlessly increasing the size of the odkfull image, because those
Python packages will be present in two different layers.

To avoid that, we forcibly remove from the odkfull staging tree any file
that is already present in the odklite staging tree.