Rename requirements.txt to requirements.txt.full.
Having a file named `requirements.txt` in the root folder leads the
mhausenblas/mkdocs-deploy-gh-pages action (that we use to deploy the
documentation) to systematically try to use that file, even though we
don't need it to build the docs (and it makes the action fail because
some of the Python modules listed in that file can't be installed in the
action's environment).

The action currently does not offer any way to opt out using the
'requirements.txt' file if it finds it, so the only workaround is not to
have a file with such a name.

closes #506