Preserve original permissions when copying template files.
When instanciating an ODK repository, copy files from the template using
`shutil.copy` instead of `shutil.copyfile`. The former has the benefit
of preserving the permissions of the source file.

For files that are generated from a Jinja2 template, explicitly transfer
the permission bits from the source template to the generated file.

close #196