Fix argument passing in the Windows run.bat script.
Windows has a peculiar way of interpreting quotes in command line
arguments: if the opening quote is not at the beginning of the argument,
then it is deemed to be part of the argument.

That is, with

  ARGUMENT='something'

the ARGUMENT variable is assigned a value of 'something', *including*
the quotes.

To make sure the quotes are only used, well, for quoting, and do not end
up in the assigned value, the opening quote must be the first character
of the argument, as in:

  'ARGUMENT=something'

This results in ARGUMENT being correctly assigned the value something,
*without* quotes.

closes #583