Fix late declaration of EDIT_PREPROCESSED.
`EDIT_PREPROCESSED` needs to be defined before the `reason_test` rule
that depends on it. Variables in targets and prerequisites are expanded
immediately when the Makefile is first parsed, so referencing a variable
in a rule before the variable is declared leads to that variable being
expanded to an empty string.