Rename IMAGE to ODK_IMAGE in the run.sh script.
In the run.sh wrapper script, rename the `IMAGE` variable to
`ODK_IMAGE`. This is for consistency with the other variables in that
script such as `ODK_DEBUG` and `ODK_JAVA_OPTS`. Likewise for `TAG`,
renamed to `ODK_TAG`.

This will allow ODK users/developers who always want to use a specific
image to safely export `ODK_IMAGE=...` and/or `ODK_TAG=...` in in their
profile scripts, without fearing that it could interfere with any other
program that might look for a `IMAGE` or `TAG` environment variable.

Likewise for the `seed-via-docker.sh` script, again for consistency.