Build `odkfull:dev` from `odklite:dev`.
Update the main Dockerfile to allow to specify a different `odklite`
version to build `odkfull` from, instead of always building from
`odklite:latest`.

Update the main Makefile so that the `build-dev` and
`publish-multiarch-dev` rules build the `odkfull:dev` image from
`odklite:dev`, so that the resulting image does contain all the latest
changes as intended.

closes #659