Add a goal to update the constraints file.
Add a script to automatically generate the constraints file by creating
a Python virtual environment, install all Python packages inside it, and
get the list of all the packages (with their versions) that were
actually installed.

That script can be called directly from the developer's machine, but it
will only produce suitable results if the developer is using a Python
version that is the same as the one provided in the ODK.

The `contraints.txt` goal in the main Makefile runs that script from
within the `obolibrary/odkbuild` image (always built on the same
Ubuntu:22.04 base as the ODK itself), to ensure the list is always
generated from the correct Python version.