Make sure the Python package `six` is present in `odklite`.
The `odklite` image is missing the Python `six` package, which is needed
by the `odk.py` script.

The problem is that the Debian meta-package `build-essential`
automatically installs `python3-six`. As a result, when we install the
Python packages when building `odkbuild` is built, `pip` does _not_
install `six`, since it is already available in
`/usr/lib/python3/dist-packages`. So the staging area in `odkbuild`
contains no trace of `six`, which is then absent from `odklite`.

The problem does not affect `odkfull` because in `odkfull`, we _also_
install the `build-essential` meta-package, which brings in
`python3-six`.

The fix is simply to explicitly install `python3-six` in the `odklite`
image.

closes #665