Source a `run.sh.conf` file if it exists.
Make the `run.sh` wrapper script check for a `run.sh.conf` file in the
`src/ontology` directory, and source that file if it exists.

This is to allow users to customize the behaviour of the wrapper script
by setting any of the ODK_* variable in that `run.sh.conf` file, instead
of having to export them in the environment or to manually set them at
each invocation of the wrapper script.

The `run.sh.conf` file is excluded from version control. It is intended
for local use only, not to be shared between all editors of the
ontology.