Run commands inside the container as the current user.
Make sure the command we run inside the ODK container is run under the
identity of the user invoking the ODK, instead of `root`.

Running as `root` will cause any file created by the process to be owned
by `root`, if the Docker daemon itself runs as a privileged user --
which happens to be the case in most GNU/Linux setups.

Running as a normal user means that we can no longer bind the calling
user's own `~/.gitconfig` file to a predictable location inside the
container (`/root/.gitconfig`), so instead we extract the informations
we need from that file in the wrapper script and pass said informations
to `odk.py` through command line arguments. The `odk.py` script will in
turn pass the informations to Git using environment variables when
initialising the repository.

closes #682