Manually add Nodesource repository.
Do not rely on a downloaded script to setup a new package source. Add
the source manually instead.

The Nodesource folks provide their OpenPGP key as as armored file only,
which needs to be dearmored to be usable by APT. We don't want to add
GnuPG to the image just to perform the dearmoring, so we store the
dearmored version straight into our repository. This has the added
benefit of protecting us from any nefarious change of the key by a
malicious actor.