Add a RDF/XML validation check.
We add a check-rdfxml script among the tools of the ODK. This script
accepts a RDF/XML file as argument and tries to parse that file using at
least the LightRDF parser, and optionally (with the `--extra` option)
with the RDFLib and the Jena parsers as well.

The standard Makefile is updated to add a new rule that runs the
aforementioned script on all RDF/XML product files. That rule is
executed as part of the `all_assets` target.