Enable optional extra RDF/XML checks.
We add a new global configuration option to the ODK:
`extra_rdfxml_checks`. If that option is enabled, then the RDF/XML
validation step will check the RDF/XML files against more parsers than
just LightRDF (for now, RDFLib and Jena).

It is disabled by default because validating with the additional parsers
is time-consuming and may be of little value. The default LightRDF
parser is already quite strict, so it is deemed unlikely that the
additional parsers would catch any problem that had not already been
caught by LightRDF. Still, the option is there for those who want/need
extra assurance that their RDF/XML files can be consumed by most
parsers.