Remove obsolete INSTALL.sh file.
The INSTALL.sh file dates from the very beginning of the ODK, back when
the ODK was still of collection of tools that needed to be “installed”.
It has never been updated and is now completely obsolete. It is also
misleading, as its mere presence can make people think that the ODK
requires an installation step.