Remove seed-via-docker-131.sh script.
This script was a temporary workaround to allow users to seed an ODK
instance with a 1.3.1 ODK image while downloading the seeding script
from the master branch of the ODK repository.

Such a workaround will no longer be needed after the 1.3.2 release. In
addition, the correct way to seed an ODK instance should be to download
the seeding script from a versioned URL, not from the master branch.