Pass environment variables to Docker in a single file.
Instead of passing environment variables to the Docker container one by
one with the `-e` option of `docker run`, put them in a temporary
`run.sh.env` file and have Docker read that file (`--env-file` option).

This is because we will soon have to pass sensitive informations (such
as GitHub tokens) as environment variables, and we don’t want to pass
them on the command line.

For the same reason, we make sure the temporary `.env` file is not
committed.