Pass GitHub token to the Docker container.
Try to pass a GitHub token as a `GH_TOKEN` environment variablle to the
Docker container, so that programs that require such a token (such as
`gh`) can be used from within a container.

The token to pass is obtained from the `GH_TOKEN` environment variable,
if such a variable is defined. Otherwise, we look for a file containing
the token at some pre-defined locations:
- .github/token
- $XDG_CONFIG_FILE/ontology-development-kit/github/token file
- $HOME/Library/Application\ Support/ontology-development-kit/github/token

The search for the token is performed after sourcing the `run.sh.conf`
file if it exists, so that file could also be used to store the token
(as a `GH_TOKEN=...` declaration).