Add safe.directory=/work to the global Git config.
If we need to run Git commands (either directly or indirectly through
another program such as `gh`) from within a container, we might run into
an issue caused by the fact the top-level directory of the repository,
as seen from within the container (`/work`), is owned by `root`. Git
flatly refuses to do anything with a repository whose top-level
directory does not belong to the calling user, even if all files and
directories below the top-level one actually belong to the right user.

We thus need to force Git to accept working on a foreign-owned
repository, by setting the `safe.directory` variable in the system-wide
Git configuration file inside the container.