Disable the LightRDF-based RDF/XML validation check.
We cannot rely on LightRDF to not incorrectly flag correct RDF/XML files
as being unparseable, so we disable the LightRDF-based check by defult.

We remove the `extra_rdfxml_checks` ODK configuration parameter (which
has never been in a public release of the ODK anyway). We only keep the
`ensure_valid_rdfxml` parameter instead, which is now OFF by default (as
it triggers the Jena-based and RDFLib-based checks, which are
time-consuming).

The `check_rdfxml` script now accept `--lightrdf`, `--rdflib`, and
`--jena` options to enable the respective checks, with RDFLib and Jena
being enabled by default (but the script itself is only called as part
of the standard workflow if the `ensure_valid_rdfxml` ODK parameter is
ON). This will allow for easy changing of the validators to used by
default in the future (including re-enabling the LightRDF validator,
if/when it is fixed).

related to #743