Make sure run.sh.env is always deleted.
To prevent people from inadvertently committing the run.sh.env file, we
make sure it is always deleted when the run.sh script terminates. We do
that by deleting the file in a EXIT trap, which is systematically
executed when the script terminates for any reason.

Not all shells execute the EXIT trap when the script termimates upon
being interrupted by a signal though, as this is not a POSIX-mandated
behaviour (Bash, Dash, and seemingly Zsh do so). For stricter
POSIX-compliant shells, we also trap the SIGINT signal, so that the
cleanup call will be executed if the user stops a long-running
invocation with Ctrl-C.