Avoid using `source` in run.sh.
The `source` shell built-in is not specified by POSIX and may not be
supported by all shells. The dash(1) shell at least, when started as
`/bin/sh`, does not recognize it. So we use instead the `.` command,
which should be supported by any POSIX-compliant shell
(https://pubs.opengroup.org/onlinepubs/009695299/utilities/dot.html).