Use a pre-built Konclude binary on arm64.
When building odkfull for x86_64, we use a binary version of Konclude
provided directly by upstream. But when building for arm64, there is no
binary available and we have to build Konclude from source (we do that
in the odkbuild image).

Building Konclude from source is time-consuming (up to several hours),
and doing that every time one builds a ODK image is a waste since
Konclude has not changed since its last release in 2021.

This commit therefore updates the arm64 build to use a custom build of
Konclude that has only been built once and for all by myself. This
considerably speeds up the build for arm64.