Allow passing configuration options to the OWL API.
Update the run.sh script to check for environment variables with a name
starting with `OWLAPI_`. All such variables are converted to OWLAPI Java
options appended to the `ODK_JAVA_OPTS` variable, which is passed to the
container.

For example, an environment variable `OWLAPI_INDENT_SIZE=2` will be
converted to
`-Dorg.semanticweb.owlapi.model.parameters.ConfigurationOptions.INDENT_SIZE=2`.

No validation of the option is performed; it is up to the user to ensure
that the name after OWLAPI_ corresponds to a valid OWLAPI configuration
parameter.

closes #778