Do not set UID/GID in seed-via-docker.sh.
PR #771 partially reverted a temporarily introduced change in the run.sh
script (the use of docker's -u option), but it did not revert a similar
change in seed-via-docker.sh. As a result, the test suite was still run
under the identity of a non-root user, which caused at least one test to
fail (the sssom test, because the sssom tool uses the bioregistry
module, which caches some data in the home directory).