Add an explicit check that the update scripts worked.
I am not entirely sure that `pip` would always return a non-zero code
upon all possible failure conditions, so just in case we explicitly
check that, if we get to the end of the script, we don't end up with an
empty constraints.txt file. Otherwise we exit with a non-zero code.