Add a Makefile rule to check if update_repo is needed.
If a fingerprint of the configuration file is available, then we store
it into the generated Makefile and we add a rule that compares the
stored fingerprint with the fingerprint of the current configuration
file as found in the repository. If the two fingerprints do not match,
this means the configuration file has been modified since the Makefile
was last generated, and therefore we print a message reminding the user
that an 'update_repo' step might be needed.