Re-enable the LightRDF-based RDF/XML check.
The ODK used to have a validation check that ensured that RDF/XML files
were valid. The check was based on LightRDF, whichwas fast enough to
make it possible to enable the check by default.

We had to disable that check because of a bug in LightRDF that caused
the library to sometimes fail to parse perfectly valid RDF/XML files
(#745). That bug has been fixed, so now we can re-enable the check by
default.

The check-rdfxml script is now invoked by default, and by default is
uses the fast LightRDF-based check. A new ODK option is added
(extra_rdfxml_checks): when enabled, it instructs the check-rdfxml
script to *also* perform the Jena- and RDFLib-based checks (which are
not enabled by default as they are more time-consuming).