Make cargo use the "sparse" protocol to download the creates index.
The default protocol that Cargo is using to download its index of
crates is somehow too resource-consuming when the ODK is built under
BuildKit (to build multiarch image). So we force Cargo to use the new
HTTP-based "sparse" protocol, which is supposedly much more efficient
and does not seem to cause issue with BuildKit.