Add SSSOM-Java to odkfull.
This commit adds SSSOM-Java to the full image: both the sssom-cli
command line tool, and the ROBOT plugin.

The command-line tool is installed so that it is readily reachable from
the PATH. The ROBOT plugin is installed in a directory dedicated to
ROBOT plugins, but *without* configuring ROBOT to look for plugins in
that directory. That is left to any workflow that require the plugin.