Automatically install ROBOT plugins.
Within an ODK workflow, ROBOT plugins should be put into the
$(TMPDIR)/plugins directory for ROBOT to find them.

This commit updates the ODK-generated Makefile to:

*) export the ROBOT_PLUGINS_DIRECTORY variable, so that ROBOT knows
where to look for plugins;
*) add a all_robot_plugins target, that automatically copies any
ODK-provided plugin (currently, only the SSSOM plugin) to the intended
plugins directory.

Therefore, rules that require the use of a plugin (be they standard
rules or rules from the custom Makefile) only need to depend on the
all_robot_plugins target.

There are two ways for a repository to add extra plugins: either by
putting them into a top-level "plugins" directory (the
"all_robot_plugins" rule will ensure that they got copied to the runtime
plugins directory in $(TMPDIR)), or by overriding the
"custom_robot_plugins" rule in the custom Makefile.