Allow refreshing the mirrors under IMP=false.
This commit overhauls the mirroring pipeline to make it independent of
the IMP variable, so that refreshing the mirrors is controlled by the
MIR variable only (except for "large" imports, for which the IMP_LARGE
variable also comes into play).

This is done using Make conditionals rather than shell conditionals.
This makes the generated Makefile much easier to read, as the entire
mirroring pipeline is simply skipped (with a single conditional at the
top) under MIR=false.