Ensure run.sh properly reports failure.
Add a 'set -e' instruction at the very beginning of the `run.sh` wrapper
script. This is to ensure the script immediately exits with a proper
error code upon the first failure.

Without such a precaution, running any workflow whose name matches the
'*update_repo*' or '*release*' patterns leads the script to
unconditionally returns a success error code, because the last command
the script executes is the final 'echo' command and _not_ the 'docker
run' command.

closes #981