Add a `run_as_root` project configuration option.
The ODK workflows now run by default under the identity of a
non-privilegied used. To run a workflow under the identity of the
super-user (which may be needed for example to install custom Debian
packages), the ODK_USER_ID variable needs either to be explicitly set to
zero or to be removed entirely from the run.sh script.

For ontologies that have custom workflows that regularly require admin
rights, it is more convenient to be able to ask once and for all that
all workflows should be run as root, directly in the -odk.yaml
configuration file, instead of having to edit the run.sh script.