Fix use of ODK_BINDS option.
The run.sh script is supposed to allow specifying extra bindings between
the local filesystem and the container, with a `ODK_BINDS` option in the
`run.sh.conf` file:

```
ODK_BINDS=/path/to/my/local/dir:/path/to/the/container/mount/point
```

This currently doesn't work when using the Docker (standard) backend
because of a single missing `g` option to the sed command that converts
the value of `ODK_BINDS` into `-v` options to pass to `docker run`.

This commit simply adds the missing `g` option.