Force system-wide installation of OBO Dashboard.
We have now officially entered the era where Python developers don't
even try anymore to maintain a coherent system, and where using virtual
environments is supposed to be the standard way to install Python
programs without breaking anything... Isn't that great?

Fuck it, we are _not_ going to create a virtual environment just for the
OBO Dashboard. If the OBO Dashboard dependencies end up breaking the
rest of our Python packages, we might as well just create a separate
Docker image just for the dashboard.