Force system-wide installation of OBO Dashboard.
Pip is now by default refusing to install anything in the system-wide
Python library path, which is considered the "private garden" of the
underlying operating system (e.g., on Ubuntu, only APT tools should add
Python packages to the system-wide path).

This may be fine in general, in a user-facing scenario, but not for the
ODK which is in effect a "read-only" system overall. The OBO Dashboard
is part of the tools/libraries we provide with the ODK and there is no
reason for us to package it separately inside a virtual environment (the
recommended way of installing non-system Python packages) -- the entire
ODK is already a "virtual environment".

Ultimately the right thing to do here would be for the OBO Dashboard
upstream to make proper releases, which we could then install at the
same time as any other Python packages.