Silence Python warning about staged installation.
PIP emits a warning when we attempt to install executable scripts in a
location that is not in the system PATH.

In this instance the warning is not warranted. The scripts are installed
in a staging location ON PURPOSE -- they are later copied over to the
final ODK images, where the scripts will end up in the system PATH.

So we shut that warning down.