run.sh: Check for spaces in the path to the current directory.
The shell wrapper script for the ODK is assuming that the directory
containing the repository does not contain spaces in its path. This is
required for the logic that binds the repository to the /work directory
inside the container to work, and we may not change that without
breaking the requirement that the wrapper script shall be POSIX
compliant (supporting space-containing paths would require using
array-typed variables, which are not specified by POSIX).

This commit adds an explicit check at the beginning of the wrapper
script to ensure that the current working directory does not contain any
spaces, and immediately abort if it does.

closes #1078