Force PIP to use setuptools < 72.
Where to begin? It's just another Python screw-up.

Developers of setuptools have announced, back in 2017 (!) their
intention to remove the `test` command (found in
setuptools.command.test) [1].

Today, they thought that 7 years of advance deprecation warnings were
enough, and they published setuptools v72.0.0 which no longer provides
that command. [2]

But the Python ecosystem being what it is (a joke that has lasted long
enough), it turns out that 7 years was not actually enough, and there
are still many Python packages out there that didn't get the memo and
that are still dependent on setuptools.command.test (including some
packages that are used in the ODK).

There are several ways of fixing this:

a) Ditch Python altogether. This would be by far the best solution, but
one that is, alas, unlikely to happen any time soon.

b) Fix _all_ the packages used by the ODK (and their dependencies) so
that they do what they should have done in the past 7 years. This is the
correct solution, but one that will take time (7 years were already not
enough, so who knows how long it is going to take?).

c) Torture the setuptools developers until they agree to revert the
removal of the `test` command. I believe this may be illegal in many
jurisdictions.

d) Force the ODK build process to use a version of `setuptools` older
than 72.0.0. This is what we are doing here. Yes, this is a local
workaround and we should favour "upstream fixing" instead (solution b),
but if we want to be be able to build the ODK before 2031 (optimistic
estimation of the time it will take to get all the Python packages
fixed), we don't have much choice.

[1] https://github.com/pypa/setuptools/issues/931
[2] https://setuptools.pypa.io/en/stable/history.html