Remove default ubuntu user account.
The Ubuntu 24.04 Docker image contains a default user account named
`ubuntu` with a UID of 1000. This is a change from the 22.04 image,
which did not contain such an account.

The presence of that account could cause some confusion as the ODK also
creates a `odkuser` user account with a default UID of 1000. To be on
the safe side, we explicitly remove the default `ubuntu` account.