Do not error out when seed is calling without a positional argument.
Calling `odk.py seed` without any argument results in a crash because
the `repo` variable (supposed to hold the name of the repository) is set
to None, which is not expected by the Jinja template (it expects a
string).

When no positional arguments are specified, we set the `repo` variable
to `noname` instead, thereby allowing the creation of a full repository
with a dummy name instead of causing an error.