Fail early if we don't have a Git username and email.
When seeding, and unless the --skipgit option has been used, we need a
username and email address to pass to Git. If we don't have those, the
seeding process will crash at the very end, after all files have been
generated.

Since the Git username and email are necessary, we should detect whether
we have them at the very beginning, and exit cleanly (with a message
telling the user what they need to to) as soon as possible, without even
initiating a process that is bound to fail.