Fix permissions issue on macOS when seeding. (#1106)
Setting Git's configuration option safe.directory to /work is enough to allow
working on a pre-seeded repository (when the repository root is bound to
/work in the container), but does not work when seeding a new
repository, because then the root of the newly created repository is not
the /work directory itself, but is in fact located two directories below
that (under /work/target/<name>). The safe.directory directive has thus
no effect when seeding.

We must set safe.directory to "*" instead, to completely disable the
permission check regardless of where the repository is. Security-wise,
this does not change anything -- that check does not bring any security
benefit in the context of the ODK, and we were already disabling it when
working on pre-seeded repositories.

closes #1105