Write $(SRCMERGED) in OFN format.
Make sure the $(SRCMERGED) intermediate file is always written in OFN
format, rather than in the same format than the -edit file.

If $(SRCMERGED) is written in OBO format, this can cause some checks
made on that file to fail because of some peculiarities of the OBO
serialisation that do not accurately reflect the contents of the
original -edit file.

closes #1107