Automatically set Java -Xmx to 90% of Docker memory.
When seeding/updating a repo, if the project does not explicitly set any
`robot_java_args`, generate code in the run.sh script to automatically
compute a default max memory setting that is dependent on the total
amount of memory that Docker is allowed to use.

The code is used if (1) we are indeed using Docker (and not Singularity)
and (2) no contradicting option is set in ODK_JAVA_OPTS (either in the
environment or in a run.sh.conf file).

closes #1074