Make sure DEFAULT_MAX_MEM is set in all conditions.
If the run.sh script is configured to use Singularity (USE_SINGULARITY
not empty), then we cannot get the max amount of memory allocated to
Singularity the same say we do for Docker (AFAIK Singularity has no such
concept as a max memory setting), but we still need to set a default
value to use for ROBOT's -Xmx. We set that value to 8G, which was the
old default value from before we added the automatic configuration.