Relax SubClassOf axioms.
When running ROBOT relax, enable the --include-subclass-of option, so
that SubClassOf axioms where the superclass is an intersection are
relaxed into equivalent separate simple SubClassOf axioms.

That behaviour is _not_ configurable, as it is deemed unnecessary:
relaxing those axioms should be uncontroversial.

In the event that someone would want to restore the previous behaviour
of _not_ relaxing SubClassOf axioms, they can do so by overriding the
RELAX_OPTIONS variable in their custom Makefile. That should be enough.

closes #1114