Fix publishing of SSSOM mapping sets.
Publishing the SSSOM mapping sets (which is not enabled by default, but
can be requested with the `release_mappings` option in the SSSOM option
group) is broken because the rsync command that should take care of
copying the mappings to the top-level directory cannot handle the case
where the input files are in a directory located above the current
directory (because of rsync's -R option).

We fix that by

(1) not including the mapping files in the RELEASE_ASSETS variables, so
that they are excluded from the rsync command;

(2) adding a new step in the `prepare_release` rule to explicitly copy
the mappings from src/mappings to the top-level mappings directory
(similarly to what we already do for the pattern files);

(3) amend the definition of the RELEASE_ASSETS_AFTER_RELEASE variable so
that it includes the mapping files in their final, top-level location.

closes #1036