Set proxy settings for Java applications.
If we detect in the environment some variables indicating that a HTTP(S)
proxy should be used, we translate them into the appropriate Java
properties so that Java applications (mostly ROBOT) inside the container
know about the proxy.

Specifically, we detect a variable like

  http_proxy=https://example.net:1234

(or HTTP_PROXY, as this variable is widely used in both uppercase and
lowercase forms) and turn it into

  -Dhttp.proxyHost=example.net
  -Dhttp.proxyPort=1234

and similarly for https_proxy (or HTTPS_PROXY), which is turned into
https.proxyHost and https.proxyPort properties.

We also detect a variable like

  no_proxy=example.com,example.org

(or, again, NO_PROXY) and turn it into

  -Dhttp.noProxyHosts=example.com|example.org

as expected by the Java virtual machine [1].

closes #1063

[1] https://docs.oracle.com/javase/8/docs/technotes/guides/net/proxies.html