Allow selective publishing of SSSOM mapping sets.
We make it possible to publish only _some_ of the SSSOM mapping sets
defined in the `sssom_mappingset_group`, by allowing each individual
product to have its own `release_mappings` flag.

The group-level flag takes precedence over the product-level flags when
it is set to True: in that case, all sets are published (as they
currently are), regardless of their individual `release_mapping` flag is
they have one.

When the group-level flag is set to False (or absent), then only the
mapping sets that have their individual `release_mapping` flag set to True
are published, if there are such sets.

Implementation-wise, to avoid cluttering the Makefile template with lots
of conditional Jinja code, the logic that determines which sets should
be published is coded in the `odk.py` script, using the `fill_missing`
feature to automatically add a constructed `released_products` slot in
the `sssom_mappingset_group`. The Jinja code in the Makefile template
can then just check whether that slot exists, and use it directly if it
does.

closes #1035