Add command to generate documentation for DOSDP patterns.
For each pipeline pattern X, we add a `dosdp-docs-X` phony target in the
Makefile to automatically generate the documentation for the patterns of
that pipeline.

This uses the (undocumented ?) `docs` subcommand of DOSDPTools.

Generated documentation is written to a new top-level directory called
`docs`, in one subdirectory `patterns/X` for each pipeline.

closes #1101