Correctly set default repo name.
If the ODK seeding configuration file does not define the name of the
repository (`repo`) and if no name is explicitly given on the command
line, we want to set a default repository name ("noname") to prevent a
crash when generating the files.

Unfortunately, the way this was done (#1097) makes any `repo` value
defined in the configuration _ignored_, because we:

* first set `repo` to "noname" if we didn't get anything from the
  command line;
* then call `load_config` with whatever value we have for `repo` (hether
  it comes from the command line or from our default), and `load_config`
  takes whatever value we give it and **overrides** any value it may
  have found in the config file.

So here, we define a default value for `project.repo` at the moment that
field is declared (as for all other default values), and do **not** set
a default value in the main function.

This way:

* if no `repo` argument is given in the command line, we get the value
  from the configuration file, or the default value if there is `repo`
  key in the config file;
* if a `repo` argument is given in the command line, it overrides both
  any value that may be set in the configuration file, and the default
  value.