Allow referring to an online repo elsewhere than on GitHub.
Add a new configuration option `repo_url` to indicate the online
location of the public repository. This is intended for repositories
that are not hosted on GitHub, where the use of github_org is therefore
inappropriate.

When generating DOSDP documentation pages, the generated documentation
will point to `repo_url` if it is present, then fallback to
`https://github.com/github_org/repo` if `github_org` is set.